requirements.txt
Django
djangorestframework
django-cors-headers
channels
channels-redis
celery==5.4.0
vine
redis
openai>=1.30.1
crawl4ai>=0.4.0
python-dotenv
Pillow
requests
beautifulsoup4
python-decouple
googlesearch-python
daphne==4.1.2
pytest-django
